+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add support for the "include note" feature, allowing views to be embedded in other notes.
 - Add `wrap` attribute setting for table views for enabling text wrapping.
 - Fix sticky table cell borders not rendering correctly when scrolling large tables in Trilium 0.47.
+- Fix attribute values not being separated correctly when there are more than two values displayed in a single table cell. Only the last two values would be separated by a space or line break instead of all values.
 
 ## 1.0.0 - 2021-04-17